下記ブランチからマージした。
svn+ssh://yutakapon@svn.sourceforge.jp/svnroot/ttssh2/branches/vs2015_warn
リビジョン6194 - 6260


git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@6286 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ teraterm/ttpmacro/ttmbuff.c
@@ -268,7 +268,7 @@ BOOL GetRawLine()
 			b = (Buff[INest])[BuffPtr[INest]];
 	}
 	LinePtr = 0;
-	LineLen = strlen(LineBuff);
+	LineLen = (WORD)strlen(LineBuff);
 	LineParsePtr = 0;
 
 	// current line number (2005.7.18 yutaka)
